backend/requirements.txt
# 파이썬 쓰는 import 추가

fastapi
uvicorn[standard]
librosa
numpy<=2.0.0
scipy
scikit-learn
pydub
yt_dlp
soundfile
faiss-cpu
speechbrain
pyyaml
huggingface-hub
torch==2.2.2
torchaudio==2.2.2
uvicorn
python-multipart
joblib
